WIP on SharedPrefs

+++ AndroidAnnotations/androidannotations/src/main/java/com/googlecode/androidannotations/annotations/sharedpreferences/DefaultInt.java
@@ -26,5 +26,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.SOURCE)
 @Target(ElementType.METHOD)
 public @interface DefaultInt {
-	long value();
+	int value();
 }
