+++ AndroidAnnotations/androidannotations/src/main/java/com/googlecode/androidannotations/annotations/sharedpreferences/DefaultLong.java
@@ -26,5 +26,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.SOURCE)
 @Target(ElementType.METHOD)
 public @interface DefaultLong {
-	int value();
+	long value();
 }
